# ENH: Improve `np.kron` performance