--- a/doc/release/upcoming_changes/21354.performance.rst
+++ b/doc/release/upcoming_changes/21354.performance.rst
@@ -1,0 +1,4 @@
+Faster ``np.kron``
+------------------
+`numpy.kron` is about 80% faster as the product is now computed
+using broadcasting.