@mattip 2022-04-17
Nice speed up. Could you add a comment about what is going on into the code?